feat: 맛집 대표 이미지 업로드 구현
- 맛집 대표 이미지 업로드
- 리팩토링
