+++ src/main/java/com/example/travelshooting/restaurant/dto/RestaurantResDto.java
@@ -15,8 +15,9 @@ public class RestaurantResDto implements BaseDtoDataType {
     private String phone;
     private String longitude;
     private String latitude;
+    private String imageUrl;
 
-    public RestaurantResDto(Long id, String region, String city, String placeName, String addressName, String roadAddressName, String phone, String longitude, String latitude) {
+    public RestaurantResDto(Long id, String region, String city, String placeName, String addressName, String roadAddressName, String phone, String longitude, String latitude, String imageUrl) {
         this.id = id;
         this.region = region;
         this.city = city;
@@ -26,5 +27,6 @@ public class RestaurantResDto implements BaseDtoDataType {
         this.phone = phone;
         this.longitude = longitude;
         this.latitude = latitude;
+        this.imageUrl = imageUrl;
     }
 }
