+++ src/main/java/com/example/travelshooting/restaurant/entity/Restaurant.java
@@ -18,16 +18,33 @@ public class Restaurant {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
+    @Column(nullable = false, length = 10)
     private String region;
+    
+    @Column(nullable = false, length = 10)
     private String city;
+
+    @Column(nullable = false, length = 30)
     private String placeName;
+
+    @Column(nullable = false, length = 50)
     private String addressName;
+
+    @Column(nullable = false, length = 50)
     private String roadAddressName;
+
+    @Column(length = 20)
     private String phone;
+
+    @Column(nullable = false, length = 100)
     private String longitude; // 경도
+
+    @Column(nullable = false, length = 100)
     private String latitude; // 위도
 
-    @OneToMany(mappedBy = "restaurant", cascade = CascadeType.ALL, orphanRemoval = true)
+    private String imageUrl;
+
+    @OneToMany(mappedBy = "restaurant")
     private List<Poster> posters = new ArrayList<>();
 
     public Restaurant(String region, String city, String placeName, String addressName, String roadAddressName, String phone, String longitude, String latitude) {
@@ -39,5 +56,9 @@ public class Restaurant {
         this.phone = phone;
         this.longitude = longitude;
         this.latitude = latitude;
+    }
+
+    public void updateImage(String imageUrl) {
+        this.imageUrl = imageUrl;
     }
 }
